pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "predictive-engine"
version = "0.1.0"
description = "Reproducible, uncertainty-aware financial prediction engine"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.6",
  "pandas>=2.2",
  "numpy>=1.26",
  "scikit-learn>=1.4",
  "jsonschema>=4.21",
  "pyyaml>=6.0",
  "joblib>=1.3",
  "pyarrow>=15.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "black>=24.0",
  "isort>=5.13",
  "mypy>=1.8",
  "ruff>=0.3",
  "types-PyYAML",
  "pandas-stubs",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
packages = ["engine"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.setuptools]
packages = ["engine"]
